High level Query method (and CreateCollection) returns RequestCharge value.
